+++ src/hooks/mutation/usePostForms.ts
@@ -27,5 +27,3 @@ const useCreateAlbaForm = () => {
 };
 
 export { useCreateAlbaForm };
-
-//router.push('/myAlbaform/owner');
