+++ src/hooks/query/useGetUser.ts
@@ -4,9 +4,11 @@
 import { fetchUser } from '@/lib/fetch/user';
 import { useQuery } from '@tanstack/react-query';
 
-export const useGetMyInfo = () => {
+export const useGetMyInfo = (enabled: boolean = true) => {
   return useQuery({
     queryKey: ['myInfo'],
     queryFn: fetchUser,
+    enabled,
+    retry: false,
   });
 };
